requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
strawberry-graphql
pytest
